Refactor api routes

+++ app/helpers/organization_helper.rb
@@ -16,7 +16,7 @@ module OrganizationHelper
   
   def meters_to_printable_km(meters)
     km = (meters / 1000.0).round(1)
-    kmi = km.to_i
+    kmi = km.to_f
     km = kmi if km == kmi
     "#{km}"
   end
